pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ah32"
version = "0.1.0"
description = "AH32 (阿蛤) - WPS + JS macro office assistant (chat + executable code)."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "Ah32 Team"}
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "python-docx>=0.8.11",
  "docx2txt>=0.8",
  "pypdf>=4.2.0",
  "langchain>=1.0.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.1.7",
  "langchain-text-splitters>=0.3.0",
  "langchainhub>=0.1.15",
  "langchain-chroma>=0.1.0",
  "langsmith>=0.1.0",
  "pydantic>=2.7.0",
  "typer>=0.12.3",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0"
]
packaging = [
  "pyinstaller>=6.0.0"
]
ppt = [
  "unstructured>=0.14.5"
]

[project.urls]
Repository = "https://codeup.aliyun.com/620a290d9d9cd9dd5cd978e0/agentic/Ah32"

[project.scripts]
ah32-server = "ah32.server.main:run"
ah32-ingest = "ah32.knowledge.ingest:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
